pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "rollbar>=1.3.0",
    "uvicorn>=0.34.2",
    "whitenoise[brotli]>=6.9.0",
]
packages = [
  { include = "task_manager" },
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "ipython>=9.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "ruff>=0.11.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]
